New Project Folder
Added new project folder with two public domain books that I found in
both English and Spanish. It's going to be harder to find free material
than I thought. One possibility is that we could make the app open to
uploading your own files. We would have to give them instructions on how
to set up the text files so that it formats properly though. And then
that could be linked to a database where they can choose to share the
content they uploaded. It could then be screened for copyright and
released to the public to use.